New methods to hide the ad banners
Added new methods to hide the ad banners and stop requesting new ads.
These methods is useful if you want to hide the ad banners at runtime,
for example if the user purchases an in-app purchase to remove ad
banners. The methods are: -hideAds on AdViewController and
AdMediationAL (if you inherit your view controller from
AdViewController, just call [self hideAds]; to hide the ad banner for
this ad). I've also added the method -hideAllAds to
AdMediationALManager: this methods call hideAds on all the allocated
instances of AdMediationAL. So if you have multiple View controllers
opened with ad banners, this method hide them all.

Note that these methods just hide the ad banners at runtime, but not
permanently (this means that the next time you start the app, or if you
instantiate a new AdMediationAL or AdViewController and call
requestAdWithDelegate, the ad banners shows again), so if you want to
disable the ad banners permanently insert a control to see if the user
purchased the "no banner" in-app purchase and if so, do not call
requestAdWithDelegate for the next time.

(Thanks to Jack Lee for the suggestion).